implement custom handlers and add documentation
