+++ beacon-chain/rpc/apimiddleware/endpoints.go
@@ -2,35 +2,44 @@ package apimiddleware
 
 import (
 	"bytes"
+	"encoding/base64"
 	"encoding/json"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"net/http"
+	"strconv"
 
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/gateway"
 )
 
-// TODO: Documentation
-func RegisterMiddlewareEndpoints() []gateway.Endpoint {
+// MiddlewareEndpointsRegistry is a registry of all endpoints that should be proxied by the API Middleware between an HTTP client and the grpc-gateway.
+//
+// All endpoints from the official Eth2 API specification must run through the middleware to maintain full compatibility with the specification.
+func MiddlewareEndpointsRegistry() []gateway.Endpoint {
 	return []gateway.Endpoint{
 		{
 			Url:         "/eth/v1/beacon/genesis",
 			GetResponse: &genesisResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/beacon/states/{state_id}/root",
 			GetResponse: &stateRootResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/beacon/states/{state_id}/fork",
 			GetResponse: &stateForkResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/beacon/states/{state_id}/finality_checkpoints",
 			GetResponse: &stateFinalityCheckpointResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:                   "/eth/v1/beacon/states/{state_id}/validators",
 			GetRequestQueryParams: []gateway.QueryParam{{Name: "id", Hex: true}, {Name: "status", Enum: true}},
@@ -40,7 +49,8 @@ func RegisterMiddlewareEndpoints() []gateway.Endpoint {
 		{
 			Url:         "/eth/v1/beacon/states/{state_id}/validators/{validator_id}",
 			GetResponse: &stateValidatorResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/beacon/states/{state_id}/validators/{validator_id}",
 			GetResponse: &stateValidatorResponseJson{},
@@ -79,15 +89,18 @@ func RegisterMiddlewareEndpoints() []gateway.Endpoint {
 		{
 			Url:         "/eth/v1/beacon/blocks/{block_id}",
 			GetResponse: &blockResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/beacon/blocks/{block_id}/root",
 			GetResponse: &blockRootResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/beacon/blocks/{block_id}/attestations",
 			GetResponse: &blockAttestationsResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:                   "/eth/v1/beacon/pool/attestations",
 			GetRequestQueryParams: []gateway.QueryParam{{Name: "slot"}, {Name: "committee_index"}},
@@ -102,25 +115,30 @@ func RegisterMiddlewareEndpoints() []gateway.Endpoint {
 			Url:         "/eth/v1/beacon/pool/attester_slashings",
 			PostRequest: &attesterSlashingJson{},
 			GetResponse: &attesterSlashingsPoolResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/beacon/pool/proposer_slashings",
 			PostRequest: &proposerSlashingJson{},
 			GetResponse: &proposerSlashingsPoolResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/beacon/pool/voluntary_exits",
 			PostRequest: &signedVoluntaryExitJson{},
 			GetResponse: &voluntaryExitsPoolResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/node/identity",
 			GetResponse: &identityResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/node/peers",
 			GetResponse: &peersResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:                   "/eth/v1/node/peers/{peer_id}",
 			GetRequestUrlLiterals: []string{"peer_id"},
@@ -130,38 +148,50 @@ func RegisterMiddlewareEndpoints() []gateway.Endpoint {
 		{
 			Url:         "/eth/v1/node/peer_count",
 			GetResponse: &peerCountResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/node/version",
 			GetResponse: &versionResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/node/syncing",
 			GetResponse: &syncingResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url: "/eth/v1/node/health",
-			Err: &gateway.DefaultErrorJson{}},
+			Err: &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/debug/beacon/states/{state_id}",
 			GetResponse: &beaconStateResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+			Hooks: gateway.HookCollection{
+				CustomHandlers: []gateway.CustomHandler{handleGetBeaconStateSsz},
+			},
+		},
 		{
 			Url:         "/eth/v1/debug/beacon/heads",
 			GetResponse: &forkChoiceHeadsResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/config/fork_schedule",
 			GetResponse: &forkScheduleResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/config/deposit_contract",
 			GetResponse: &depositContractResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 		{
 			Url:         "/eth/v1/config/spec",
 			GetResponse: &specResponseJson{},
-			Err:         &gateway.DefaultErrorJson{}},
+			Err:         &gateway.DefaultErrorJson{},
+		},
 	}
 }
 
@@ -190,6 +220,97 @@ func prepareGraffiti(endpoint gateway.Endpoint, _ http.ResponseWriter, _ *http.R
 	if block, ok := endpoint.PostRequest.(*beaconBlockContainerJson); ok {
 		b := bytesutil.ToBytes32([]byte(block.Message.Body.Graffiti))
 		block.Message.Body.Graffiti = hexutil.Encode(b[:])
+	}
+	return nil
+}
+
+func handleGetBeaconStateSsz(m *gateway.ApiProxyMiddleware, endpoint gateway.Endpoint, writer http.ResponseWriter, request *http.Request) (handled bool) {
+	if !beaconStateSszRequested(request) {
+		return false
+	}
+
+	if errJson := prepareRequestForProxying(m, endpoint, request); errJson != nil {
+		gateway.WriteError(writer, errJson, nil)
+		return
+	}
+	grpcResponse, errJson := gateway.ProxyRequest(request)
+	if errJson != nil {
+		gateway.WriteError(writer, errJson, nil)
+		return
+	}
+	grpcResponseBody, errJson := gateway.ReadGrpcResponseBody(grpcResponse.Body)
+	if errJson != nil {
+		gateway.WriteError(writer, errJson, nil)
+		return
+	}
+	if errJson := gateway.DeserializeGrpcResponseBodyIntoErrorJson(endpoint.Err, grpcResponseBody); errJson != nil {
+		gateway.WriteError(writer, errJson, nil)
+		return
+	}
+	if endpoint.Err.Msg() != "" {
+		gateway.HandleGrpcResponseError(endpoint.Err, grpcResponse, writer)
+		return
+	}
+	responseJson := &beaconStateSszResponseJson{}
+	if errJson := gateway.DeserializeGrpcResponseBodyIntoContainer(grpcResponseBody, responseJson); errJson != nil {
+		gateway.WriteError(writer, errJson, nil)
+		return
+	}
+	responseSsz, errJson := serializeMiddlewareResponseIntoSsz(responseJson.Data)
+	if errJson != nil {
+		gateway.WriteError(writer, errJson, nil)
+		return
+	}
+	if errJson := writeMiddlewareResponseHeaderAndBody(grpcResponse, responseSsz, writer); errJson != nil {
+		gateway.WriteError(writer, errJson, nil)
+		return
+	}
+	if errJson := gateway.Cleanup(grpcResponse.Body); errJson != nil {
+		gateway.WriteError(writer, errJson, nil)
+		return
+	}
+
+	return true
+}
+
+func beaconStateSszRequested(request *http.Request) bool {
+	return request.Header["Accept"][0] == "application/octet-stream"
+}
+
+func prepareRequestForProxying(m *gateway.ApiProxyMiddleware, endpoint gateway.Endpoint, request *http.Request) gateway.ErrorJson {
+	request.URL.Scheme = "http"
+	request.URL.Host = m.GatewayAddress
+	request.RequestURI = ""
+	request.URL.Path = "/eth/v1/debug/beacon/states/{state_id}/ssz"
+	if errJson := gateway.HandleUrlParameters(endpoint.Url, request, []string{}); errJson != nil {
+		return errJson
+	}
+	return nil
+}
+
+func serializeMiddlewareResponseIntoSsz(data string) (sszResponse []byte, errJson gateway.ErrorJson) {
+	// Serialize the SSZ part of the deserialized value.
+	b, err := base64.StdEncoding.DecodeString(data)
+	if err != nil {
+		e := fmt.Errorf("could not decode response body into base64: %w", err)
+		return nil, &gateway.DefaultErrorJson{Message: e.Error(), Code: http.StatusInternalServerError}
+	}
+	return b, nil
+}
+
+func writeMiddlewareResponseHeaderAndBody(grpcResponse *http.Response, responseSsz []byte, writer http.ResponseWriter) gateway.ErrorJson {
+	for h, vs := range grpcResponse.Header {
+		for _, v := range vs {
+			writer.Header().Set(h, v)
+		}
+	}
+	writer.Header().Set("Content-Length", strconv.Itoa(len(responseSsz)))
+	writer.Header().Set("Content-Type", "application/octet-stream")
+	writer.Header().Set("Content-Disposition", "attachment; filename=beacon_state_ssz")
+	writer.WriteHeader(grpcResponse.StatusCode)
+	if _, err := io.Copy(writer, ioutil.NopCloser(bytes.NewReader(responseSsz))); err != nil {
+		e := fmt.Errorf("could not write response message: %w", err)
+		return &gateway.DefaultErrorJson{Message: e.Error(), Code: http.StatusInternalServerError}
 	}
 	return nil
 }
